refactor: exceptions 디렉토리 생성
